+++ main.tf
@@ -12,4 +12,11 @@ resource "cloudflare_dns_record" "this" {
   content  = var.content
   priority = var.type == "MX" ? var.priority : null
   proxied  = var.proxied
+
+  lifecycle {
+    precondition {
+      condition     = !endswith(var.name, data.cloudflare_zone.current.name) && !endswith(var.name, ".")
+      error_message = "The name of the DNS record cannot end with the zone name or a dot."
+    }
+  }
 }