+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to
 
 ### Added
 
+- **BREAKING**: Validate that the record name (`name`) does not end with the DNS
+  zone name or a dot.
 - Validate that the name of the record (`name`) is not empty.
 - Validate that the `priority` input variable is greater than `0` when the
   record type is `MX`.